+++ tests/testthat/test_tune.R
@@ -26,6 +26,27 @@ test_that("tune function works without measure", {
   expect_measure(instance$objective$measures[[1]])
 })
 
+test_that("tune interface is equal to TuningInstanceSingleCrit", {
+  tune_args = formalArgs(tune)
+  tune_args = tune_args[tune_args != "method" & tune_args != "..."]
+  tune_args[tune_args == "measures"] = "measure"
+
+  instance_args = formalArgs(TuningInstanceSingleCrit$public_methods$initialize)
+  instance_args = c(instance_args, "term_evals", "term_time")
+
+  expect_set_equal(tune_args, instance_args)
+})
+
+test_that("tune interface is equal to TuningInstanceMultiCrit", {
+  tune_args = formalArgs(tune)
+  tune_args = tune_args[tune_args != "method" & tune_args != "..."]
+
+  instance_args = formalArgs(TuningInstanceMultiCrit$public_methods$initialize)
+  instance_args = c(instance_args, "term_evals", "term_time")
+
+  expect_set_equal(tune_args, instance_args)
+})
+
 # evaluate_default ------------------------------------------------------------------
 
 test_that("evaluate_default works", {